Adds test for TroveUpdated event
Added the TroveManagerTester contract that allows the test to set the base rate. The base rate normally changes
with redemptions and issuance to prevent bank runs. Refactored the tests to set the tester version of a contract
in the fixture rather than having the tests reference contracts.musdTester it just calls contracts.musd. The
test then checks for the existence of an tester function before calling it and fails the test if it doesnt exist.

Added a helper function to search for specific events in the transaction results.