Adds SortedTroves insert trove logic
Added test into BorrowerOperations to check that the users position is added into the list of SortedTroves.
This required adding the troves data into the SortedTroves contract which is basically a doubly linkedlist
ordering the positions by collateralisation ratio. As all of the troves have the same collateral type the
price doesnt matter only the debt to collateral ratio.

The insert function is restricted so that only the BorrowerOperations or TroveManager contracts can call it.
After the contracts are created the params for the SortedTroves contract must set the addresses for
BorrowerOperations and TroveManager before renouncing ownership of the contract.

Whale balances where added to the users for testing.