Adds Liquidate and close trove
Ported the liquidate trove code so that the test to make sure that the adjustments to the trove managers collateral
and debt levels work correctly and dont impact the next trove thats opened. Ported the close and repay trove code to
ensure that after a trove is closed the borrower can still open another position.

Porting tests for

Borrowing at zero base rate changes the PCV contract MUSD fees collected
Borrowing at non-zero base rate sends requested amount to the user
Reverts when opening the trove would cause the TCR of the system to fall below the CCR

The tests have also be refactored into groups.