Standardize fixture setup for tests.
Using different fixtures across tests files results in a InvalidSnapshotError when the usage goes
fixture1, fixture2, fixture1 in the files being executed. When the tests are executed individually
they execute without error.

PriceFeedTestnet.sol is deleted and replaced with MockAggregator so that the fixture just loads
the PriceFeed.sol contract and then has it load the MockAggregator.