Port over the `provideToSP` tests for `StabilityPool`
This is *heavily* refactored from the thusd repo.

The original tests were very bloated, often making assertions that seemed totally unrelated to the description. We're following a philosophy where each test does the minimal amount of setup to test what it says it's testing, and then *just* tests that.

This removes a lot of assertions. I think this is fine - if those code paths end up being totally uncovered, our coverage tool should pick it up later and we'll patch it over with more simple tests.