Restructure interest updating calls
We update the system (and trove) interest at the top of all publically callable
functions, and remove all other (now useless) calls.

Along the way, we make sure that every time we do that, we're also updating the
default pool.

This squashes a few hard-to-find potential bugs around not detecting that we're
in recovery mode.