TokenDeployer contract for MUSD
The TokenDeployer allows the governance to deploy MUSD token to the
chain using a stable address. The TokenDeployer contract should be
deployed earlier to the chain using the EIP2470 singleton factory.

The TokenDeployer has not been wired up yet with the deploy process -
expect it to happen in the next commits.