Separate deployer from governance address in TokenDeployer
The deployer address is EOA and governance is a multisig. Using EOA will
allow to perform the deployment process in one shot, without waiting for
the governance multisig to assemble and schedule a transaction. The
control over the deployed token is passed to the governance as a part of
the deployment so the only thing that can go wrong in case deployer EOA
is compormised are the contract addresses which can be updated later on
the token by the governance.