Provide mainnet RPC URL as an environment variable
We do not want to make RPC URLs public before the launch and there are
some temporary issues with them as partners are setting them up. It is
much more convenient to provide it as an environment variable now.

Also, I fixed the typo in how Ethereum RPC URL is read from env
variable.