Publish just deployment artifacts to NPM
We extract the artifacts of deployed contracts to a separate package so
that we can import them as a minimal dependency in other projects.

The NPM contract packages we used to publish contained all the
contract's source code, tests, deployment scripts, and a long list of
solidity development-related dependencies. This was way too heavy,
especially since all we need are just ABIs and addresses of deployed
contracts.

The default Hardhat directory structure is modified by this change:
- The contracts compilation artifacts are moved from the `artifacts/` to
  the `build/` directory.
- The deployment artifacts are moved from the `deployments/` to the
  `artifacts/deployments/` directory.
- The `artifacts/` directory is now becoming a home for the
  `@mezo-org/musd-contracts` package.

Publishing from the `solidity/` directory is no longer possible given
the `private: true` flag in `package.json`. Publishing from
`solidity/artifacts` will publish a package with just artifacts to the
NPM registry under `@mezo-org/musd-contracts`.